+++ lib/libesp32/berry/default/be_modtab.c
@@ -83,6 +83,11 @@ be_extern_native_module(haspmota);
 #endif // USE_LVGL_HASPMOTA
 #endif // USE_LVGL
 #ifdef USE_MATTER_DEVICE
+#ifdef USE_WS2812
+#ifdef USE_BERRY_ANIMATION
+be_extern_native_module(animation);
+#endif // USE_BERRY_ANIMATION
+#endif // USE_WS2812
 be_extern_native_module(matter);
 #endif // USE_MATTER_DEVICE
 
@@ -217,6 +222,11 @@ BERRY_LOCAL const bntvmodule_t* const be_module_table[] = {
 #ifdef USE_MATTER_DEVICE
     &be_native_module(matter),
 #endif // USE_MATTER_DEVICE
+#ifdef USE_WS2812
+#ifdef USE_BERRY_ANIMATION
+    &be_native_module(animation),
+#endif // USE_BERRY_ANIMATION
+#endif // USE_WS2812
 #endif // TASMOTA
     CUSTOM_NATIVE_MODULES
     /* user-defined modules register end */