+++ lib/libesp32/berry_animation/src/berry_animation.h
@@ -1,0 +1,8 @@
+// force include of module by including this file
+
+#ifndef __BERRY_ANIMATION__
+#define __BERRY_ANIMATION__
+
+
+
+#endif // __BERRY_ANIMATION__